# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API - Contateiros",
        "description": "Documentação gerada com L5 Swagger",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/usuario": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Lista todos os usuários",
                "description": "Retorna todos os usuários cadastrados no sistema.",
                "operationId": "974e17e11effd01817c102768172d2c8",
                "responses": {
                    "200": {
                        "description": "Usuários encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "usuarios": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "teste"
                                                    },
                                                    "usuario": {
                                                        "type": "string",
                                                        "example": "teste"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "teste@teste.com"
                                                    },
                                                    "telefone": {
                                                        "type": "string",
                                                        "example": "+5544991087686"
                                                    },
                                                    "id_funcao": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_secao": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "ativo": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-01T03:14:07.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-01T03:14:07.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum usuário encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nenhum usuário encontrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Cria um novo usuário",
                "operationId": "ac5cce9f3789a6701b6f22e24e58e9ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "usuario",
                                    "email",
                                    "telefone",
                                    "id_funcao",
                                    "id_secao",
                                    "ativo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "usuario": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "telefone": {
                                        "type": "string"
                                    },
                                    "id_funcao": {
                                        "type": "integer"
                                    },
                                    "id_secao": {
                                        "type": "integer"
                                    },
                                    "ativo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "usuario": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nome": {
                                                    "type": "string"
                                                },
                                                "usuario": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "telefone": {
                                                    "type": "string"
                                                },
                                                "id_funcao": {
                                                    "type": "integer"
                                                },
                                                "id_secao": {
                                                    "type": "integer"
                                                },
                                                "ativo": {
                                                    "type": "boolean"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erros de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O usuário deve ser informado."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}